Create ADC10_UART.ino
Timer 2 generating 32 KHz clock to trigger overflow interrupt every 31.25 microSeconds.   Interrupt Service Routine increments a counter and compares it to 640.  If equal, Read the 10 bit Analog to Digital Conversion result.  The result is passed to a function that converts a 16 bit unsigned integer to a character array representing the integer.  The characters in the array are passed to the USART for serial transmission to the serial monitor or plotter on a PC.   The Interrupt Service Routine also has a counter that creates and alternating delay of .0375 S on time and .75 S off time to blink the LED on Pin 13.  
The upper 8 bits are output to an LED display.  This is an exercise in fast efficient programming by accessing registers directly without using libraries and avoiding the use of the processor and memory intensive Divide and Modulo functions.  This version of the program is 1020 bytes.  The Example Blink program that comes with the Arduino environment is 928 bytes.