Create arduinoADC_USART_inlineASM.ino
Demonstrating the use of inline asm() to set Pin 2 at the beginning of the Interrupt routing and clear Pin 2 a the end of the Interrupt routing.  This facilitates timing the service routine using an oscilloscope to measure the pulse duration.  Testing the current version of the file reveals that 7 % of the processor time is used processing the interrupt service routine.